requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
Pillow==10.0.1
mysqlclient==2.1.1
redis==5.0.1
celery==5.3.4
django-celery-results==2.5.1
django-redis==5.3.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0